+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to
 - 🐛(front) fix target blank links in chat #103
 - 🚑️(posthog) pass str instead of UUID for user PK #134
 - ⚡️(web-search) keep running when tool call fails #137
+- ✨(summarize): new summarize tool integration #78
+
+### Removed
+
+- 🔥(posthog) remove posthog middleware for async mode fix #146
 
 
 ## [0.0.7] - 2025-10-28